% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\title{Add mcintegration objects}
\usage{
add(x, y)
}
\arguments{
\item{x}{An mcintegration object.}

\item{y}{An mcintegration object.}
}
\value{
Returns a numeric value corresponding to the result of x added with
the result of y.
}
\description{
Add mcintegration objects
}
\examples{
x <- mcintegrator(sin, 0, pi / 2)
y <- mcintegrator(function(x) {-x}, -1, 0)
add(x, y)
}
