% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrator.R
\name{integrator}
\alias{integrator}
\title{Integrates a function using a Monte Carlo}
\usage{
integrator(fn, a, b)
}
\arguments{
\item{fn}{A univariate function that is non-negative and monotone in the interval [a;b].}

\item{a}{Start of integration interval.}

\item{b}{End of integration interval.}
}
\value{
Returns a list with the following values:
\itemize{
\item \code{result}: The approximation of the integral.
\item \code{point}: A dataframe with two columns x and y, specifying all the point
coordinates used in the approximation.
\item \code{points_above}: A dataframe like \code{points} containing the points above
the function line.
\item \code{points_below}: A dataframe like 'points_above', but with the
points below the function line.
\item \code{interval}: A vector of two numeric values specifying the integration interval.
\item \code{upper}: The largest value of the function in the integration interval.
\item \code{fineness}: The number of points used for the approximation.
\item \code{func}: The integrand.
}
}
\description{
Integrates a function using a Monte Carlo
}
\examples{
x <- integrator(sin, 0, pi / 2)
y <- integrator(function(x) {-x}, -1, 0)
}
