% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcintegrator.R
\name{mcintegrator}
\alias{mcintegrator}
\title{Definite integration of non-negative monotone functions using Monte Carlo}
\usage{
mcintegrator(fn, a, b)
}
\arguments{
\item{fn}{A univariate function that is non-negative and monotone in the interval [a;b].}

\item{a}{Start of integration interval.}

\item{b}{End of integration interval.}
}
\value{
Returns an object with class mcintegrator. See
?integrator for details.
}
\description{
Definite integration of non-negative monotone functions using Monte Carlo
}
\examples{
x <- mcintegrator(sin, 0, pi / 2)
y <- mcintegrator(function(x) {-x}, -1, 0)
}
