% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{plot.mcintegrator}
\alias{plot.mcintegrator}
\title{Plot mcintegrator object}
\usage{
\method{plot}{mcintegrator}(x, ...)
}
\arguments{
\item{x}{An mcintegrator object.}

\item{...}{Extra arguments read by plot}
}
\value{
A plot of the function integrated, with points under the function
line in the color blue, and points above the function line in the
color pink.
}
\description{
Plot mcintegrator object
}
\examples{
x <- mcintegrator(sin, 0, pi / 2)
plot(x)
}
