% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{print.mcintegrator}
\alias{print.mcintegrator}
\title{Print mcintegrator object}
\usage{
\method{print}{mcintegrator}(x, ...)
}
\arguments{
\item{x}{An mcintegrator object.}

\item{...}{Extra arguments read by print}
}
\value{
Prints the approximation of the integral.
}
\description{
Print mcintegrator object
}
\examples{
x <- mcintegrator(sin, 0, pi / 2)
print(x)
}
