% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtract.R
\name{subtract}
\alias{subtract}
\title{Subtract mcintegration objects}
\usage{
subtract(x, y)
}
\arguments{
\item{x}{An mcintegration object.}

\item{y}{An mcintegration object.}
}
\value{
Returns a numeric value corresponding to the result of y subtracted
from the result of x.
}
\description{
Subtract mcintegration objects
}
\examples{
x <- mcintegrator(sin, 0, pi / 2)
y <- mcintegrator(function(x) {-x}, -1, 0)
subtract(x, y)
}
