% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{summary.mcintegrator}
\alias{summary.mcintegrator}
\title{Summarize mcintegrator object}
\usage{
\method{summary}{mcintegrator}(object, ...)
}
\arguments{
\item{object}{An mcintegrator object.}

\item{...}{Extra arguments read by summary}
}
\value{
Prints a summary containing result, fineness, and integration interval
of the mcintegrator object.
}
\description{
Summarize mcintegrator object
}
\examples{
x <- mcintegrator(sin, 0, pi / 2)
summary(x)
}
